Arvados-DCO-1.1-Signed-off-by: Peter Amstutz <peter.amstutz@curii.com>

+++ services/api/app/controllers/sys_controller.rb
@@ -19,16 +19,23 @@ class SysController < ApplicationController
         in_batches(of: 15).
         update_all('is_trashed = true')
 
-      # Sweep trashed projects and their contents (as well as role
-      # groups that were trashed before #18340 when that was
-      # disallowed)
+      # Want to make sure the #update_trash hook on the Group class
+      # runs.  It does a couple of important things:
+      #
+      # - For projects, puts all the subprojects in the trashed_groups table.
+      #
+      # - For role groups, outbound permissions are deleted.
       Group.
-        where('delete_at is not null and delete_at < statement_timestamp()').each do |project|
-          delete_project_and_contents(project.uuid)
+        where("is_trashed = false and trash_at < statement_timestamp()").each do |grp|
+        grp.is_trashed = true
+        grp.save
       end
+
+      # Sweep groups and their contents that are ready to be deleted
       Group.
-        where('is_trashed = false and trash_at < statement_timestamp()').
-        update_all('is_trashed = true')
+        where('delete_at is not null and delete_at < statement_timestamp()').each do |group|
+          delete_project_and_contents(group.uuid)
+      end
 
       # Sweep expired tokens
       ActiveRecord::Base.connection.execute("DELETE from api_client_authorizations where expires_at <= statement_timestamp()")
@@ -43,19 +50,21 @@ class SysController < ApplicationController
     if !p
       raise "can't sweep group '#{p_uuid}', it may not exist"
     end
-    # First delete sub projects
-    Group.where({group_class: 'project', owner_uuid: p_uuid}).each do |sub_project|
-      delete_project_and_contents(sub_project.uuid)
-    end
-    # Next, iterate over all tables which have owner_uuid fields, with some
-    # exceptions, and delete records owned by this project
-    skipped_classes = ['Group', 'User']
-    ActiveRecord::Base.descendants.reject(&:abstract_class?).each do |klass|
-      if !skipped_classes.include?(klass.name) && klass.columns.collect(&:name).include?('owner_uuid')
-        klass.where({owner_uuid: p_uuid}).in_batches(of: 15).destroy_all
+    if p.group_class == 'project'
+      # First delete sub projects and owned filter groups
+      Group.where({owner_uuid: p_uuid}).each do |sub_project|
+        delete_project_and_contents(sub_project.uuid)
+      end
+      # Next, iterate over all tables which have owner_uuid fields, with some
+      # exceptions, and delete records owned by this project
+      skipped_classes = ['Group', 'User']
+      ActiveRecord::Base.descendants.reject(&:abstract_class?).each do |klass|
+        if !skipped_classes.include?(klass.name) && klass.columns.collect(&:name).include?('owner_uuid')
+          klass.where({owner_uuid: p_uuid}).in_batches(of: 15).destroy_all
+        end
       end
     end
-    # Finally delete the project itself
+    # Finally delete the group itself
     p.destroy
   end
 end
